Create Fixtures from Snapshot

+++ priv/repo/migrations/20190117193815_create_auction_fixture_record.exs
@@ -8,7 +8,7 @@ defmodule Oceanconnect.Repo.Migrations.CreateAuctionFixtureRecord do
       add :vessel_id, references(:vessels, on_delete: :nothing)
       add :fuel_id, references(:fuels, on_delete: :nothing)
 
-      add :price, :integer
+      add :price, :decimal
       add :quantity, :integer
       add :eta, :utc_datetime_usec
       add :etd, :utc_datetime_usec
@@ -18,7 +18,7 @@ defmodule Oceanconnect.Repo.Migrations.CreateAuctionFixtureRecord do
       add :original_vessel_id, references(:vessels, on_delete: :nothing)
       add :original_fuel_id, references(:fuels, on_delete: :nothing)
 
-      add :original_price, :integer
+      add :original_price, :decimal
       add :original_quantity, :integer
       add :original_eta, :utc_datetime_usec
       add :original_etd, :utc_datetime_usec
